fix: #57 벡엔드 models 분리, 라우터 수정 등등
